backend/auth_service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
asyncpg==0.28.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
email-validator==2.0.0
bcrypt==4.0.1
python-dotenv==1.0.0
httpx==0.25.0
pytest==7.4.2
pytest-asyncio==0.21.1
python-jose==3.3.0
PyJWT